Revision and complete cleanup
The iteration over each line has been revisited. Now, instead of looking for the next vertex line, we move onto the next line, check which type of line it is (vertex or face or other), then apply the appropriate conversion to that line.

To do : finish the conversion of vertex postfixes, create a conversion for faces, convert all to integers instead of characters/strings, convert to binary, write to file.